requirements.txt
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.7.0
bqplot==0.12.40
branca==0.6.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colour==0.1.5
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
Django==4.2.4
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
earthengine-api==0.1.364
eerepr==0.0.4
executing==1.2.0
filelock==3.12.2
flake8==6.1.0
folium==0.14.0
fonttools==4.42.1
future==0.18.3
gdown==4.7.1
geemap==0.25.0
geocoder==1.38.1
geojson==3.0.1
geospatial==0.0.2
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
importlib-resources==6.0.1
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipyleaflet==0.17.3
ipython==8.14.0
ipytree==0.2.2
ipywidgets==8.1.0
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.17.3
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
leafmap==0.23.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.25.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
plotly==5.16.1
prompt-toolkit==3.0.39
protobuf==4.24.1
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.3
pyshp==2.3.1
PySocks==1.7.1
pystac==1.8.3
pystac-client==0.7.3
python-box==7.0.1
python-dateutil==2.8.2
pytz==2023.3
ratelim==0.1.6
requests==2.31.0
rsa==4.9
scooby==0.7.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
tenacity==8.2.3
tomli==2.0.1
tqdm==4.66.1
traitlets==5.9.0
traittypes==0.2.1
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
whitebox==2.3.1
whiteboxgui==2.3.0
widgetsnbextension==4.0.8
xarray==2023.8.0
xyzservices==2023.7.0
zipp==3.16.2
